requirements.txt
openai>=1.40
chromadb>=0.5
lxml>=5.0
pydantic>=2.0
tiktoken>=0.7
python-dotenv>=1.0
jupyter>=1.0
ipywidgets>=8.0
pytest>=8.0
matplotlib>=3.8
